+++ app.py
@@ -5,8 +5,7 @@ import os
 import openai
 from dotenv import load_dotenv
 from colorama import Fore, Back, Style
-
-
+from utils import get_response
 
 load_dotenv()
 
@@ -22,77 +21,45 @@ Engage in a professional and informative manner, and avoid using jargon or compl
 Remember, your purpose is to assist users with their healthcare-related inquiries and provide educational support. With these guidelines in mind, proceed to help users with their questions, concerns, and medicine recommendations.
 """
 
-TEMPERATURE = 0.5
-MAX_TOKENS = 500
-FREQUENCY_PENALTY = 0
-PRESENCE_PENALTY = 0.6
-# limits how many questions we include in the prompt
-MAX_CONTEXT_QUESTIONS = 10
+# TEMPERATURE = 0.5
+# MAX_TOKENS = 500
+# FREQUENCY_PENALTY = 0
+# PRESENCE_PENALTY = 0.6
+# # limits how many questions we include in the prompt
+# MAX_CONTEXT_QUESTIONS = 100
 
-def get_response(instructions, previous_questions_and_answers, new_question):
-    """Get a response from ChatCompletion
+# def get_response(instructions, previous_questions_and_answers, new_question):
+#     """Get a response from ChatCompletion
 
-    Args:
-        instructions: The instructions for the chat bot - this determines how it will behave
-        previous_questions_and_answers: Chat history
-        new_question: The new question to ask the bot
+#     Args:
+#         instructions: The instructions for the chat bot - this determines how it will behave
+#         previous_questions_and_answers: Chat history
+#         new_question: The new question to ask the bot
 
-    Returns:
-        The response text
-    """
-    # build the messages
-    messages = [
-        { "role": "system", "content": instructions },
-    ]
-    # add the previous questions and answers
-    for question, answer in previous_questions_and_answers[-MAX_CONTEXT_QUESTIONS:]:
-        messages.append({ "role": "user", "content": question })
-        messages.append({ "role": "assistant", "content": answer })
-    # add the new question
-    messages.append({ "role": "user", "content": new_question })
+#     Returns:
+#         The response text
+#     """
+#     # build the messages
+#     messages = [
+#         { "role": "system", "content": instructions },
+#     ]
+#     # add the previous questions and answers
+#     for question, answer in previous_questions_and_answers[-MAX_CONTEXT_QUESTIONS:]:
+#         messages.append({ "role": "user", "content": question })
+#         messages.append({ "role": "assistant", "content": answer })
+#     # add the new question
+#     messages.append({ "role": "user", "content": new_question })
 
-    completion = openai.ChatCompletion.create(
-        model="gpt-3.5-turbo",
-        messages=messages,
-        temperature=TEMPERATURE,
-        max_tokens=MAX_TOKENS,
-        top_p=1,
-        frequency_penalty=FREQUENCY_PENALTY,
-        presence_penalty=PRESENCE_PENALTY,
-    )
-    return completion.choices[0].message.content
-
-
-def get_moderation(question):
-    """
-    Check the question is safe to ask the model
-
-    Parameters:
-        question (str): The question to check
-
-    Returns a list of errors if the question is not safe, otherwise returns None
-    """
-
-    errors = {
-        "hate": "Content that expresses, incites, or promotes hate based on race, gender, ethnicity, religion, nationality, sexual orientation, disability status, or caste.",
-        "hate/threatening": "Hateful content that also includes violence or serious harm towards the targeted group.",
-        "self-harm": "Content that promotes, encourages, or depicts acts of self-harm, such as suicide, cutting, and eating disorders.",
-        "sexual": "Content meant to arouse sexual excitement, such as the description of sexual activity, or that promotes sexual services (excluding sex education and wellness).",
-        "sexual/minors": "Sexual content that includes an individual who is under 18 years old.",
-        "violence": "Content that promotes or glorifies violence or celebrates the suffering or humiliation of others.",
-        "violence/graphic": "Violent content that depicts death, violence, or serious physical injury in extreme graphic detail.",
-        "medical error":"System: I strive to provide accurate medical information, but I may occasionally make mistakes. If you notice any errors or have concerns about my responses, please let me know.Please remember that I am an AI assistant and not a substitute for professional medical advice. Always consult a healthcare professional for personalized recommendations and treatment plans.Your feedback is valuable in helping me improve. Thank you for understanding."
-    }
-    response = openai.Moderation.create(input=question)
-    if response.results[0].flagged:
-        # get the categories that are flagged and generate a message
-        result = [
-            error
-            for category, error in errors.items()
-            if response.results[0].categories[category]
-        ]
-        return result
-    return None
+#     completion = openai.ChatCompletion.create(
+#         model="gpt-3.5-turbo",
+#         messages=messages,
+#         temperature=TEMPERATURE,
+#         max_tokens=MAX_TOKENS,
+#         top_p=1,
+#         frequency_penalty=FREQUENCY_PENALTY,
+#         presence_penalty=PRESENCE_PENALTY,
+#     )
+#     return completion.choices[0].message.content
 from flask import Flask
 
 app = Flask(__name__, template_folder='templates')
@@ -129,7 +96,7 @@ def chat():
         previous_questions_and_answers.append((new_question, response))
 
         # print the response
-        return ("Here you go: ",response)
+        return {"response": response}
         
 if __name__ == "__main__":
     app.run(debug=True)
